==> 0. Explore Bot-Adversarial-Dialogue classification.ipynb <==
import pandas as pd

train = pd.read_csv('bot_adversarial_dialogue_datasets_with_persona/train.txt', sep='\t', header=None)

train.head()

train.shape

print(train[0].loc[1])

for i, row in train.iterrows():
    if i == 100:
        break
        
    personas = row[4].split('\\n')
    for p in personas:
        print(p)
    text = row[0].strip('text:')
    utterances = text.split('\\n')
    for u in utterances:
        print(u)
    print(row[3])
    print(row[1])
    print('------------------------------------------')
