==> 1. Get DialoGPT to continue Reddit conversations.ipynb <==
# !pip install ipywidgets

from transformers import AutoModelForCausalLM, AutoTokenizer
import torch


tokenizer = AutoTokenizer.from_pretrained("microsoft/DialoGPT-small")
model = AutoModelForCausalLM.from_pretrained("microsoft/DialoGPT-small")

# Let's chat for 5 lines
for step in range(5):
    # encode the new user input, add the eos_token and return a tensor in Pytorch
    new_user_input_text = input(">> User:") + tokenizer.eos_token
    new_user_input_ids = tokenizer.encode(new_user_input_text, return_tensors='pt')
    print('input text:', new_user_input_text)
    print('input token ids:', new_user_input_ids)

    # append the new user input tokens to the chat history
    bot_input_ids = torch.cat([chat_history_ids, new_user_input_ids], dim=-1) if step > 0 else new_user_input_ids
    print('bot input ids:', bot_input_ids)

    # generated a response while limiting the total chat history to 1000 tokens, 
    chat_history_ids = model.generate(bot_input_ids, max_length=1000, pad_token_id=tokenizer.eos_token_id)
    
    # pretty print last ouput tokens from bot
    print("History + DialoGPT: {}".format(tokenizer.decode(chat_history_ids[:,:][0], skip_special_tokens=False)))
    print("DialoGPT: {}".format(tokenizer.decode(chat_history_ids[:, bot_input_ids.shape[-1]:][0], skip_special_tokens=True)))


import pandas as pd
pd.set_option('display.max_colwidth', None)

from transformers import AutoModelForCausalLM, AutoTokenizer
import torch

device = "cuda:0" if torch.cuda.is_available() else "cpu"

model = AutoModelForCausalLM.from_pretrained("microsoft/DialoGPT-small").to(device)
tokenizer = AutoTokenizer.from_pretrained("microsoft/DialoGPT-small")

conv_df = pd.read_csv('out/conv/2010-11.tsv', header=None, delimiter = '\t').sample(500, random_state = 42)
conv_df.columns = ['id', 'context', 'response']

import re 
from nltk.tokenize import TweetTokenizer

def preprocess_text(
    txt,
    dataset_eos_token = 'EOS',
    tokenizer_eos_token = '<|endoftext|>'
):
    # remove "title : " prefixes
    if txt[:8] == 'title : ':
        txt = txt[8:]
    
    txt = str(txt).lower()
    
    # url and tag
    words = []
    for word in txt.split():
        if word[0] == '#': # don't allow tag
            continue
        i = word.lower().find('http')
        if i >= 0:
            word = word[:i] + ' ' + '__url__'
        words.append(word.strip())
    txt = ' '.join(words)

    # remove illegal char
    txt = txt.replace(chr(92),'') # chr(92) = '\'. as twitter has 'b\/c' rather than 'b/c'
    txt = txt.replace("b/c","because").replace('j/k','just kidding').replace('w/o','without').replace('w/','with')
    txt = re.sub('__mention__','MENTION',txt)
    txt = re.sub('__url__','URL',txt)
    txt = re.sub(r"[^A-Za-z0-9()\[\]:,.!?'“” ]", " ", txt)
    txt = re.sub('MENTION','__mention__',txt)
    txt = re.sub('URL','__url__',txt)

    tokenizer = TweetTokenizer(preserve_case=True)
    txt = ' ' + ' '.join(tokenizer.tokenize(txt)) + ' '
    
    # remove un-necessary space
    txt = ' '.join(txt.split())
    
    # replace 'EOS' with tokenizer's EOS
    txt_utterance_split = txt.split(dataset_eos_token.lower())
    txt = tokenizer_eos_token.join([s.strip() for s in txt_utterance_split])
    
    return txt



prepro = preprocess_text("title : I thought the Reddit Admins Promised no more ads with auto play sounds ? ? ? EOS You're supposed to block ads . EOS Not for sites you like . EOS Yes , all sites . Everywhere .")
print(prepro)
tokenizer(prepro)

%load_ext autoreload
%autoreload 2

from dataset import DialogLMDataset
from torch.utils.data import DataLoader
from collate_fns import DialogCollate, DialogCollateExperimental

# train sets
dataset_train = DialogLMDataset(
    df = conv_df,
    src_col = 'context',
    src_eos_token = ' EOS ',
    tokenizer = tokenizer,
    tgt_col = 'response',
)
print(dataset_train[-1])

tokenizer.decode(dataset_train[-1]['src'])

# collate object
collate_fn = DialogCollateExperimental(
    tokenizer = tokenizer,
    max_len = 512,
    _targets_ignore_index = -100,
    _pad_token_id = tokenizer.eos_token_id  # 0
)

# dataloader
loader_val = DataLoader(
    dataset_train,
    batch_size = 1,
    shuffle = False,
    collate_fn = collate_fn
)

# example usage
for i, batch in enumerate(loader_val):
    if i == 3:
        break

example_id = 0
print('input_ids:\n',       batch['input_ids'][example_id])
print('position_ids:\n',    batch['position_ids'][example_id])
print('token_type_ids:\n',  batch['token_type_ids'][example_id])
print('target_ids:\n',      batch['target_ids'][example_id])
print('attention_masks:\n', batch['attention_masks'][example_id])

tokenizer.decode(batch['input_ids'][example_id])

import random

# generate responses to each example in training set

# perhaps it's most convenient to have the random cutoffs during the dataset instantiation 
# and do a few runs through the dataset to generate random stuff

n_samples_to_generate = 100
n_valid_samples_so_far = 0
samples = {
    'context': [],
    'response': []
}
index_of_example = 0
epoch = 0  # how many times have we made a pass through the data?
while n_valid_samples_so_far < n_samples_to_generate:
    
    print(f'\rProcessing sample {n_valid_samples_so_far} of {n_samples_to_generate}', end='', flush=True)
    
    # grab the example
    context = conv_df.iloc[index_of_example]['context'] + ' EOS ' + conv_df.iloc[index_of_example]['response']
    context = preprocess_text(context)
    context_split_utterances = [s.strip() for s in context.split('<|endoftext|>')]
    
    # # randomly cut off conversation at one of the turns
    # cutoff_point = random.choice(range(len(context_split_utterances)+1))
    # context_split_utterances = context_split_utterances[:cutoff_point]
    # if len(context_split_utterances) == 0:  # do unconditional generation
    #     context = ''
    #     input_ids = None
    #     len_input_ids = 0
    # else:
    #     context = '<|endoftext|>'.join(context_split_utterances) + '<|endoftext|>'
    #     input_ids = tokenizer(context, add_special_tokens=False, return_tensors="pt")["input_ids"].to(device)
    #     len_input_ids = input_ids.shape[-1]
    
    # model continues conversation
    input_ids = tokenizer(context, add_special_tokens=False, return_tensors="pt")["input_ids"].to(device)
    outputs = model.generate(
        inputs = input_ids, 
        max_length = 1000,
        pad_token_id = tokenizer.eos_token_id,
        do_sample = True,  # do_sample = True; otherwise all questions will be identical
        top_p = 0.95,      # nucleus sampling
        top_k = 0          # deactivate top-k words sampling
    ).cpu()
#     print('\n',input_ids)
#     break
#     print(outputs)
#     print(outputs[0])
    output = outputs[0]
#     print(len(output));break
    output_bot_only = output[len(input_ids[0]):]
#     print(output_bot_only)
    output_bot_only_decoded = tokenizer.decode(output_bot_only, skip_special_tokens=True)
    print(f'\n{context} *** {output_bot_only_decoded}')
    # validate; skip if invalid
    output_preproc = preprocess_text(output_bot_only_decoded).strip()
    if len(output_preproc) == 0:  # put validation condition here
        index_of_example = (index_of_example + 1) % len(conv_df)
        print("\n^Error")
        continue
    
    # reformat and save
    if context[-13:] == '<|endoftext|>':
        context = context[:-13]
    context_reformat = context.split('<|endoftext|>')
    context_reformat = ' EOS '.join(context_reformat)
    response_reformat = output_preproc.strip('<|endoftext|>')
    samples['context'].append(context_reformat)
    samples['response'].append(response_reformat)
    
    # dump every 1000 samples
    if n_valid_samples_so_far % 10 == 0:  # dump every 1000 prompts
        print(f"\nSaving {len(samples['context'])} samples to csv")
        pd.DataFrame(samples).to_csv(f'red_lm_prompts/zero_shot/prompts_{int(n_valid_samples_so_far/10)-1}0_{int(n_valid_samples_so_far/10)}0.csv', index=False)
        samples = {'context': [], 'response': []}
    
    # increment example
    if index_of_example == len(conv_df)-1:
        epoch += 1
    index_of_example = (index_of_example + 1) % len(conv_df)
    n_valid_samples_so_far += 1
