==> 3. Formatting Dialogs for ParlAI Classifier.ipynb <==
import pandas as pd
import glob

test = pd.read_csv('bot_adversarial_dialogue_datasets_with_persona/test.txt', sep='\t', header=None)

response_df

def construct_parlai_dialog(utterances_list):
    return 'text:' + '\\n'.join([str(u).replace("'", '').replace('"', '') for u in utterances_list])
    
    
files = glob.glob('synthetic_data_responses/combined_responses.csv')
for f_num, f in enumerate(files):
    
    response_df = pd.read_csv(f)
    response_df.columns = ['prompt', 'response']
    example_ids_suffix = f.split('\\')[-1].strip('.csv').strip('response_')
    print(f'Processing csv {f_num} of {len(files)}, filename {f}')
    
    texts = []
    speakers_to_eval = []
    for i, row in response_df.iterrows():

        # initialize dialog with the prompt
        texts.append(  construct_parlai_dialog(row['prompt'].strip('title : ').split(' EOS '))  )
        speakers_to_eval.append('speaker_to_eval:human')

        # add bot response
        texts.append(  construct_parlai_dialog(row['prompt'].strip('title : ').split(' EOS ') + [row['response']])  )
        speakers_to_eval.append('speaker_to_eval:bot')

    print()
    responses_parlai_df = pd.DataFrame({
        'text': texts,
        'labels': 'labels:__ok__',
        'episode_done': 'episode_done:True',
        'speaker_to_eval': speakers_to_eval,
    })
    responses_parlai_df.to_csv(f'synthetic_data_responses/combined_responses_formatted.csv', index=False, header=False, sep='\t')
