==> 8. Training the Improved Model using Synthetic Offensive Data.ipynb <==
%load_ext autoreload
%autoreload 2

from dataset import DialogLMDataset
from torch.utils.data import DataLoader
from collate_fns import DialogCollate, DialogCollateExperimental
from models import NegativePositiveTrainingLM
from torch.utils.data import DataLoader, Dataset, RandomSampler, SequentialSampler


import torch
from transformers import AutoConfig, AutoTokenizer
from sklearn.model_selection import train_test_split

# utils
from data_utils import InfiniteDataGen
from logging_utils import create_json, update_json, load_json

import os
import json
import numpy as np
import pandas as pd
pd.set_option('display.max_colwidth', None)
import datetime
import random

df = pd.read_csv('data/reddit_conversations/censored_reddit_data.csv', index_col=0)
df.columns = ['context', 'response', 'original_response']
df = df.dropna(subset='context')
df['context'] = df['context'].astype(str)
df['response'] = df['response'].astype(str)
df['original_response'] = df['original_response'].astype(str)

extra_df = pd.read_csv('offensive_synthetic_convs_censored.csv')
extra_df.columns = ['context', 'response', 'original_response']
extra_df = extra_df.dropna(subset='context')
extra_df['context'] = extra_df['context'].astype(str)
extra_df['response'] = extra_df['response'].astype(str)
extra_df['original_response'] = extra_df['original_response'].astype(str)

n_train = 120_000
df_train, df_val = df.iloc[:n_train], df.iloc[n_train:]
print(df_train.shape, df_val.shape)
print(extra_df.shape)

df_train = pd.concat([df_train, extra_df]).reset_index().drop('index', axis=1)
df_train

df_train.to_csv('augmented_training_data.csv', index=False)

# tokenizer
tokenizer = AutoTokenizer.from_pretrained(
    'microsoft/DialoGPT-small'
)

# train sets
dataset_train = DialogLMDataset(
    df = df_train, #.iloc[:10],
    src_col = 'context',
    src_eos_token = ' EOS ',
    tokenizer = tokenizer,
    tgt_col = 'response'
)

# val sets
dataset_val = DialogLMDataset(
    df = df_val,
    src_col = 'context',
    src_eos_token = ' EOS ',
    tokenizer = tokenizer,
    tgt_col = 'response'
)
print()
print(dataset_val[3])
tokenizer.decode(dataset_val[-1]['src'])

# example dataset and dataloader usage

# # collate object
# collate_fn = DialogCollateExperimental(
#     tokenizer = tokenizer,
#     max_len = 512,
#     _targets_ignore_index = -100,
#     _pad_token_id = tokenizer.eos_token_id  # 0
# )
collate_fn = DialogCollate(
    tokenizer = tokenizer,
    max_len = 512,
    _targets_ignore_index = -100,
    _pad_token_id = tokenizer.eos_token_id  # 0
)

# dataloader
loader_train = DataLoader(
    dataset_train,
    batch_size = 2,
    shuffle = False,
    collate_fn = collate_fn
)
loader_val = DataLoader(
    dataset_val,
    batch_size = 2,
    shuffle = False,
    collate_fn = collate_fn
)

# example usage
for batch in loader_train:
    break

example_id = 0
print('input_ids:\n',       batch['input_ids'][example_id])
# print('position_ids:\n',    batch['position_ids'][example_id])
# print('token_type_ids:\n',  batch['token_type_ids'][example_id])
print('target_ids:\n',      batch['target_ids'][example_id])
# print('attention_masks:\n', batch['attention_masks'][example_id])

tokenizer.decode(batch['input_ids'][example_id])

Explanation of target IDs: `-100` indicates that this position will not contribute to the loss. This behavior is specified by setting the `ignore_index` option in the `CrossEntropy` loss function, which by default takes the value `-100`. In order to decode the targets, we replace `-100`'s arbitrarily with `0`'s, and the ID `0` happens to correspond to the exclamation mark.

def train_one_positive_epoch(
    pos_loader, model_wrapper
):
    model_wrapper.model.train()

    # loop through the adversarial, negative batches
    # repeat until all negative samples have low enough log probability under the model, 
    # or a maximum number of iterations has occurred
    batch_wise_pos_losses  = []
    for pos_batch_id, pos_batch in enumerate(pos_loader):
        
        # model_wrapper.optimizer_pos.zero_grad()
        # results = model_wrapper.positive_step_basic(pos_batch, debug=True)
        results = model_wrapper.positive_step(pos_batch, debug=False)
        
        batch_wise_pos_losses.append(results['loss'].item())
        
        print(f'\rProcessing pos batch {pos_batch_id} of {len(pos_loader)}, loss {results["loss"].item():.6f}', end='', flush=True)


    return {
        'batch_wise_pos_losses': batch_wise_pos_losses,
        'average_pos_loss': np.array(batch_wise_pos_losses).mean(),
    }

CONFIG = {
    'EXPERIMENT_NAME': 'improvement_with_synthetic_data',
    
    'PERFORM_NEGATIVE_TRAINING': False,
    
    'GRAD_ACCUM_STEPS': 1,
    'START_ITER': 0,
    'TRAIN_ITERS': 3,
    'POSITIVE_RATIO': 1,
    
    'LOGGING_DIR': './logs/',
    'MODEL_SAVE_DIR': './checkpoints/',
    
    'HUGGINGFACE_MODEL_NAME': 'microsoft/DialoGPT-small',  # 'microsoft/DialoGPT-medium'
    'POS_LR': 5e-5,
    'NEG_LR': 5e-5,
    'UPDATES_PER_BATCH': 20,
    
    'EXAMPLE_WEIGHT_MODE': 'decay',
    'EXAMPLE_WEIGHT_CARE_MODE': 'sample_avg',
    'EXAMPLE_WEIGHT_REJECTION_THRESHOLD': -7.0,
    
    'SEED': 0
}

# training
    
# preparing model checkpointing
if not os.path.exists(os.path.join(CONFIG['MODEL_SAVE_DIR'], CONFIG['EXPERIMENT_NAME'])): 
    os.makedirs(os.path.join(CONFIG['MODEL_SAVE_DIR'], CONFIG['EXPERIMENT_NAME']))

# data loaders
train_sampler = RandomSampler(dataset_train)
loader_train = DataLoader(
    dataset_train,
    sampler = train_sampler,
    batch_size = 2,
    # shuffle = True,
    collate_fn = collate_fn,
    drop_last = True
)
eval_sampler = SequentialSampler(dataset_val)
loader_val = DataLoader(
    dataset_val,
    sampler = eval_sampler,
    batch_size = 2,
    # shuffle = False,
    collate_fn = collate_fn,
    drop_last = True
)

# model
model_config = AutoConfig.from_pretrained(CONFIG['HUGGINGFACE_MODEL_NAME'])
gradient_accumulation_steps = 1
model_wrapper = NegativePositiveTrainingLM(
    model_name = CONFIG['HUGGINGFACE_MODEL_NAME'],
    model_config = model_config,
    pos_lr = CONFIG['POS_LR'], 
    neg_lr = CONFIG['NEG_LR'],
    num_opt_steps = len(loader_train) // CONFIG['GRAD_ACCUM_STEPS'] * CONFIG['TRAIN_ITERS'],
    tokenizer = tokenizer,
    device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
)

# evaluate before training
# reproducibility
random.seed(CONFIG['SEED'])
np.random.seed(CONFIG['SEED'])
torch.manual_seed(CONFIG['SEED'])
torch.cuda.manual_seed_all(CONFIG['SEED'])
# # train_loss = model_wrapper.perplexity_on_dataset(loader_train)
# val_loss   = model_wrapper.perplexity_on_dataset(loader_val)
# # print('\nTrain perplexity:', train_loss)
# print('\nVal perplexity:', val_loss)

# train_loss = model_wrapper.nll_on_dataset(loader_train)
val_loss   = model_wrapper.nll_on_dataset(loader_val)
# print('\nTrain NLL:', train_loss)
print('\nVal NLL:', val_loss)

# restart the optimizer at each pos-neg iteration
model_wrapper.reset_optimizers()

# zero-grad the model before training
model_wrapper.model.zero_grad()

# reproducibility
random.seed(CONFIG['SEED'])
np.random.seed(CONFIG['SEED'])
torch.manual_seed(CONFIG['SEED'])
torch.cuda.manual_seed_all(CONFIG['SEED'])
    
# detect previous progress here
epoch_wise_neg_losses = []
epoch_wise_pos_losses = []
batch_wise_neg_losses = []
batch_wise_pos_losses = []

# train for `CONFIG['TRAIN_ITERS']` "negative epochs"
# one "negative epoch" = one pass through all batches in the entire negative dataset:
# for each negative batch, do (one negative update + POSITIVE_RATIO positive updates)*20
# notice the difference with usual training: here we do 20 gradient updates for each batch in the negative dataloader
for it in range(CONFIG['START_ITER'], CONFIG['START_ITER'] + CONFIG['TRAIN_ITERS']):  

    # # restart the optimizer at each pos-neg iteration
    # model_wrapper.reset_optimizers()
    
    # train on one epoch
    res = train_one_positive_epoch(
        pos_loader    = loader_train, 
        model_wrapper = model_wrapper
    )
        
    # evaluate val set
    # train_loss = model_wrapper.perplexity_on_dataset(loader_train)
    # val_loss   = model_wrapper.perplexity_on_dataset(loader_val)
    train_loss = 0 # model_wrapper.nll_on_dataset(loader_train)
    val_loss   = model_wrapper.nll_on_dataset(loader_val)
    print('\nTrain loss:', train_loss)
    print('\nVal loss:', val_loss)
    
    # record losses
    epoch_wise_neg_losses.append(res.get('average_neg_loss', None))
    epoch_wise_pos_losses.append(res.get('average_pos_loss', None))
    batch_wise_neg_losses.extend(res.get('batch_wise_neg_losses', []))
    batch_wise_pos_losses.extend(res.get('batch_wise_pos_losses', []))
    
    # log results
    log_dict = {
        'config': CONFIG,
        'log_datetime': datetime.datetime.now().strftime('%Y-%m-%d %H:%M:%S'),
        'num_negative_epochs': it,
        'train_neg_loss': res.get('average_neg_loss', None),
        'train_pos_loss': res.get('average_pos_loss', None),
        'train_num_updates_per_batch': res.get('batch_wise_num_updates', None),
        'train_epoch_wise_neg_losses': epoch_wise_neg_losses,
        'train_epoch_wise_pos_losses': epoch_wise_pos_losses,
        'train_batch_wise_neg_losses': batch_wise_neg_losses,
        'train_batch_wise_pos_losses': batch_wise_pos_losses,
    }
    create_json(
        target_dir = os.path.join(CONFIG['LOGGING_DIR'], CONFIG['EXPERIMENT_NAME']), 
        filename = f'epoch-{it}_trainloss-{res.get("average_pos_loss", np.nan):.6f}_valloss-{val_loss:.6f}.json',
        dict_to_save = log_dict
    )
    
    # save model
    model_wrapper.save_model_and_optimizers(
        info_dict = log_dict,  # additional information you want to log
        save_path = os.path.join(
            CONFIG["MODEL_SAVE_DIR"], 
            CONFIG["EXPERIMENT_NAME"], 
            f'epoch-{it}_trainloss-{res.get("average_pos_loss", np.nan):.6f}_valloss-{val_loss:.6f}'
        )
    )

# which experiment to load?
EXPERIMENT_TO_LOAD = 'improvement_with_synthetic_data'
CHECKPOINT_TO_LOAD = 'epoch-2_trainloss-2.113416_valloss-2.624816'
LOGGING_DIR        = './logs/'
MODEL_SAVE_DIR     = './checkpoints/'

# load config
log_dict = load_json(
    target_dir = os.path.join(LOGGING_DIR, EXPERIMENT_TO_LOAD),
    filename   = CHECKPOINT_TO_LOAD + '.json'
)

# grab the config from the logs
CONFIG = log_dict['config']

# model
model_config = AutoConfig.from_pretrained(CONFIG['HUGGINGFACE_MODEL_NAME'])
gradient_accumulation_steps = 1
model_wrapper = NegativePositiveTrainingLM(
    model_name = CONFIG['HUGGINGFACE_MODEL_NAME'],
    model_config = model_config,
    pos_lr = CONFIG['POS_LR'], 
    neg_lr = CONFIG['NEG_LR'],
    num_opt_steps = len(loader_train) // CONFIG['GRAD_ACCUM_STEPS'] * CONFIG['TRAIN_ITERS'],
    tokenizer = tokenizer,
    device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
)

# load
model_wrapper.load_model(
    checkpoint_root = CONFIG['MODEL_SAVE_DIR'],
    experiment_name = CONFIG['EXPERIMENT_NAME'],
    checkpoint_name = CHECKPOINT_TO_LOAD,
)

# Chat

df_train.sample()

device = "cuda:0" if torch.cuda.is_available() else "cpu"

# initial conversation starter
init_conv = """
Not funny . Never has been , never will be . Period .
"""

# preprocess to correct format
utterances = init_conv.split('EOS')
utterances = [s.strip() for s in utterances]
init_conv_formatted = tokenizer.eos_token.join(utterances)

# Let's chat for 5 lines
for step in range(5):
    
    ###########################################
    # start with some initial text
    if step == 0:
        new_user_input_ids = tokenizer.encode(init_conv_formatted + tokenizer.eos_token, return_tensors='pt').to(device)
        print('>> User:')
        for u in init_conv_formatted.split(tokenizer.eos_token):
            print(u)
    else:
        # encode the new user input, add the eos_token and return a tensor in Pytorch
        new_user_input_ids = tokenizer.encode(input(">> User:") + tokenizer.eos_token, return_tensors='pt').to(device)
    # print(new_user_input_ids)
    ###########################################
    # # or start from scratch
    # new_user_input_ids = tokenizer.encode(input(">> User:") + tokenizer.eos_token, return_tensors='pt').to(device)
    # ###########################################

    # append the new user input tokens to the chat history
    bot_input_ids = torch.cat([chat_history_ids, new_user_input_ids], dim=-1) if step > 0 else new_user_input_ids
    print(bot_input_ids)

    # generated a response while limiting the total chat history to 1000 tokens, 
    chat_history_ids = model_wrapper.model.generate(
        bot_input_ids, max_length=200,
        pad_token_id=tokenizer.eos_token_id,  
        no_repeat_ngram_size=3,       
        do_sample=True, 
        top_k=100, 
        top_p=0.7,
        temperature=0.8
    )
    
    # pretty print last ouput tokens from bot
    print("RickBot: {}".format(tokenizer.decode(chat_history_ids[:, bot_input_ids.shape[-1]:][0], skip_special_tokens=True)))

device = "cuda:0" if torch.cuda.is_available() else "cpu"

# Let's chat for 5 lines
for step in range(5):
    # encode the new user input, add the eos_token and return a tensor in Pytorch
    new_user_input_ids = tokenizer.encode(input(">> User:") + tokenizer.eos_token, return_tensors='pt').to(device)
    # print(new_user_input_ids)

    # append the new user input tokens to the chat history
    bot_input_ids = torch.cat([chat_history_ids, new_user_input_ids], dim=-1) if step > 0 else new_user_input_ids

    # generated a response while limiting the total chat history to 1000 tokens, 
    chat_history_ids = model_wrapper.model.generate(
        bot_input_ids, max_length=200,
        pad_token_id=tokenizer.eos_token_id,  
        no_repeat_ngram_size=3,       
        do_sample=True, 
        top_k=100, 
        top_p=0.7,
        temperature=0.8
    )
    
    # pretty print last ouput tokens from bot
    print("RickBot: {}".format(tokenizer.decode(chat_history_ids[:, bot_input_ids.shape[-1]:][0], skip_special_tokens=True)))

device = "cuda:0" if torch.cuda.is_available() else "cpu"

# Let's chat for 5 lines
for step in range(5):
    # encode the new user input, add the eos_token and return a tensor in Pytorch
    new_user_input_ids = tokenizer.encode(input(">> User:") + tokenizer.eos_token, return_tensors='pt').to(device)
    # print(new_user_input_ids)

    # append the new user input tokens to the chat history
    bot_input_ids = torch.cat([chat_history_ids, new_user_input_ids], dim=-1) if step > 0 else new_user_input_ids

    # generated a response while limiting the total chat history to 1000 tokens, 
    chat_history_ids = model_wrapper.model.generate(
        bot_input_ids, max_length=200,
        pad_token_id=tokenizer.eos_token_id,  
        no_repeat_ngram_size=3,       
        do_sample=True, 
        top_k=100, 
        top_p=0.7,
        temperature=0.8
    )
    
    # pretty print last ouput tokens from bot
    print("RickBot: {}".format(tokenizer.decode(chat_history_ids[:, bot_input_ids.shape[-1]:][0], skip_special_tokens=True)))

device = "cuda:0" if torch.cuda.is_available() else "cpu"

# Let's chat for 5 lines
for step in range(5):
    # encode the new user input, add the eos_token and return a tensor in Pytorch
    new_user_input_ids = tokenizer.encode(input(">> User:") + tokenizer.eos_token, return_tensors='pt').to(device)
    # print(new_user_input_ids)

    # append the new user input tokens to the chat history
    bot_input_ids = torch.cat([chat_history_ids, new_user_input_ids], dim=-1) if step > 0 else new_user_input_ids

    # generated a response while limiting the total chat history to 1000 tokens, 
    chat_history_ids = model_wrapper.model.generate(
        bot_input_ids, max_length=200,
        pad_token_id=tokenizer.eos_token_id,  
        no_repeat_ngram_size=3,       
        do_sample=True, 
        top_k=100, 
        top_p=0.7,
        temperature=0.8
    )
    
    # pretty print last ouput tokens from bot
    print("RickBot: {}".format(tokenizer.decode(chat_history_ids[:, bot_input_ids.shape[-1]:][0], skip_special_tokens=True)))

# Responses to validation set

df_train

df_val

prompt = ["What do you think about Elon Musk?<|endoftext|>"] * 1  # generate a batch of 10 independent samples
inputs = tokenizer(prompt, return_tensors="pt")["input_ids"].to(model_wrapper.device)
prompt_length = len(tokenizer.decode(inputs[0]))

outputs = model_wrapper.model.generate(
    inputs, 
    max_length=200,
    pad_token_id=tokenizer.eos_token_id,  
#     no_repeat_ngram_size=3,       
    do_sample=True, 
    top_k=100, 
    top_p=0.7,
    temperature = 0.8
).cpu()
decoded_outputs = [tokenizer.decode(output) for output in outputs]

decoded_outputs

loader = DataLoader(
    dataset_val,
    batch_size = 1,
    shuffle = False,
    collate_fn = collate_fn
)
num_batches_to_show = 100

responses_dict = {}
example_id = 0
for i, batch in enumerate(loader):
    break

print(batch)
tokenizer.decode(batch['input_ids'][example_id])

# each tensor to device
input_ids       = batch['input_ids'].to(self.device)
# position_ids    = batch['position_ids'].to(self.device)
# token_type_ids  = batch['token_type_ids'].to(self.device)
# target_ids      = batch['target_ids'].to(self.device)
attention_mask  = batch['attention_masks'].to(self.device)

outputs = self.model.generate(
    inputs = input_ids, 
    max_new_tokens = max_new_tokens,
    pad_token_id = self.tokenizer.eos_token_id,
    do_sample = True,  # do_sample = True; otherwise all questions will be identical
    top_p = 0.95,      # nucleus sampling
    top_k = 0,         # deactivate top-k words sampling
    attention_mask = attention_mask  # do not pay attention to padding
).cpu()

decoded_outputs = [self.tokenizer.decode(output) for output in outputs]
