==> 9c. Analyzing the proposed model.ipynb <==
import pandas as pd
import glob

dfs = []
for filename in glob.glob('no_safety_responses/*.csv'):
    df = pd.read_csv(filename, index_col=None)
    dfs.append(df)
    
all_df = pd.concat(dfs).reset_index().drop('index', axis=1)
all_df

all_df.to_csv('human_data_only_responses/combined_human_data_only_responses.csv', index=False)

all_df.iloc[:10000].to_csv('human_data_only_responses/combined_human_data_only_responses_10k.csv', index=False)

scores_df = pd.read_csv('synthetic_data_responses/combined_responses_formatted_scored.tsv', sep='\t', header=None)
scores_df.columns = ['text', 'is_offensive_true', 'episode_done', 'speaker_to_eval', 'is_offensive_pred']
scores_df = scores_df[['text', 'episode_done', 'speaker_to_eval', 'is_offensive_pred']]

scores_df

# Percent offensive prompts

prompts_df = scores_df[scores_df['speaker_to_eval'] == 'speaker_to_eval:human']
num_examples = len(prompts_df)
num_offensive = len(prompts_df[prompts_df['is_offensive_pred'] == '__notok__'])
offensive_prompts_df = scores_df[
      (scores_df['speaker_to_eval'] == 'speaker_to_eval:human')  # 'human' here refers to the red LM
    & (scores_df['is_offensive_pred'] == '__notok__')
].copy()
num_offensive/num_examples

offensive_prompts_df

# how many unique prompts?
offensive_prompts_df.drop_duplicates(subset='text').iloc[3]['text']

offensive_prompts_df.iloc[-6]['text']

offensive_prompts_df.sample()['text'].iloc[0]

# Percent offensive replies

replies_df = scores_df[scores_df['speaker_to_eval'] == 'speaker_to_eval:bot']
num_examples  = len(replies_df)
num_offensive = len(replies_df[replies_df['is_offensive_pred'] == '__notok__'])
offensive_replies_df = scores_df[
      (scores_df['speaker_to_eval'] == 'speaker_to_eval:bot')  # 'bot' here refers to the target LM
    & (scores_df['is_offensive_pred'] == '__notok__')
].copy()
num_offensive/num_examples

offensive_replies_df

offensive_replies_df.iloc[0]['text']

offensive_replies_df.sample().iloc[0]['text']

# how many with non-sequitur responses?
num_ok = 0
for i, row in scores_df.iterrows():
    response = row['text'].split('\\n')[-1]
    if response[:45] == 'Hey do you want to talk about something else?':
        num_ok += 1

num_ok

# this is not the same number as when searching through the whole dataset.
# that means the offensiveness classifier thought some non-sequiturs were not offensive, but some were. WTF.

a = 'Hey do you want to talk about something else?'
len(a)

2266-640

(1626-1403)/1626

# Censoring negative responses

import json
with open('data/reddit_conversations/parlai_formatted/topic_splits.json') as json_file:
    topic_splits = json.load(json_file)

x_topics = topic_splits['train']

import random
random.choice(x_topics)

scores_df

import random

def nonsequitur_response(topic):
    return f'Hey do you want to talk about something else? How about we talk about {topic}?'

prompts = []
responses = []
original_responses = []
num_examples = len(scores_df)
for i, row in scores_df.iterrows():
    
    if i % 1000 == 0:
        print(f'\rProcessing example {i} of {num_examples}', end='', flush=True)
    
    text = row['text'].strip('text:')
    
    # split between prompt and reponse
    utterances = text.split('\\n')
    prompt = ' EOS '.join(utterances[:-1])
    response = utterances[-1]
    
    # replace response with non-sequitur if offensive
    if row['speaker_to_eval'] == 'speaker_to_eval:bot' and row['is_offensive_pred'] == '__notok__':
        censored_response = nonsequitur_response(random.choice(x_topics))
    else:
        censored_response = response
        
    # save
    prompts.append(prompt)
    responses.append(censored_response)
    original_responses.append(response)
    
censored_df = pd.DataFrame({
    'prompt': prompts,
    'response': responses,
    'original_response': original_responses
})

censored_df.loc[5]['response'], censored_df.loc[5]['original_response'] 

censored_df.to_csv('data/reddit_conversations/censored_reddit_data.csv')

censored_df
